some stream group by funcs ex added

+++ hello-lambda/src/com/ailhanli/ex9/AdvanceStream.java
@@ -7,15 +7,15 @@ import java.util.stream.IntStream;
 public class AdvanceStream {
 
 	public static void main(String[] args) {
-		System.out.println(Arrays.asList(1, 2, 3, 4, 5).stream().collect(Collectors.summingInt(a -> a)));
+		System.out.println(Arrays.asList(1, 2, 3, 4, 5).stream().collect(Collectors.summingInt(a -> a)).intValue());
 
-		System.out.println(Arrays.asList(1, 2, 3, 4, 5).stream().collect(Collectors.averagingInt(x -> x)));
-		
-		System.out.println(Arrays.asList(1,2,3,4,5).stream().collect(Collectors.maxBy(Integer::compareTo)).get());
-		
-		Arrays.asList(1,2,3,4,5).stream().max(Integer::compareTo);
-		
-		System.out.println(IntStream.rangeClosed(1, 100).boxed().collect(Collectors.summingDouble(a -> a)));
-		
+		System.out.println(Arrays.asList(1, 2, 3, 4, 5).stream().collect(Collectors.averagingInt(x -> x)).intValue());
+
+		System.out.println(Arrays.asList(1, 2, 3, 4, 5).stream().collect(Collectors.maxBy(Integer::compareTo)).get());
+
+		Arrays.asList(1, 2, 3, 4, 5).stream().max(Integer::compareTo);
+
+		System.out
+				.println(IntStream.rangeClosed(1, 100).boxed().collect(Collectors.summingDouble(a -> a)).doubleValue());
 	}
 }
